lambdas/poller/requirements.txt
python-twitter
requests
boto3